refactored : separated logic and view
